requirements.txt
django
psycopg2-binary
python-dotenv
django-bootstrap5